- minor cpp style changes - fixes file directory as commented by Cardozo
