+++ catkin_home/src/arm_vision/include/IncludersCloud.h
@@ -1,31 +1,27 @@
 #include <ros/ros.h>
 #include <sensor_msgs/PointCloud2.h>
+
 #include <iostream>
-#include<iterator> // for iterators 
-#include <vector>  // for vectors 
+#include <iterator>  // for iterators
+#include <vector>    // for vectors
 //#include<conio.h>
-#include<math.h>
+#include <math.h>
 //#include <pcl/ros/conversions.h>
 #include <pcl_conversions/pcl_conversions.h>
 //#include <pcl/conversions.h>
+#include <my_pcl_tutorial/SegmentedClustersArray.h>
+#include <pcl/ModelCoefficients.h>
+#include <pcl/filters/extract_indices.h>
 #include <pcl/filters/passthrough.h>
 #include <pcl/filters/voxel_grid.h>
-#include <pcl/filters/extract_indices.h>
-
-
-#include <pcl/point_cloud.h>
-#include <pcl/point_types.h>
-
-#include <pcl/ModelCoefficients.h>
 #include <pcl/io/pcd_io.h>
+#include <pcl/kdtree/kdtree.h>
+#include <pcl/point_cloud.h>
 #include <pcl/point_types.h>
 #include <pcl/sample_consensus/method_types.h>
 #include <pcl/sample_consensus/model_types.h>
-#include <pcl/segmentation/sac_segmentation.h>
-
-#include <pcl/kdtree/kdtree.h>
 #include <pcl/segmentation/extract_clusters.h>
-#include <my_pcl_tutorial/SegmentedClustersArray.h>
+#include <pcl/segmentation/sac_segmentation.h>
 //#include <obj_recognition/ClusterData.h>
 
 // PCL specific includes
