+++ catkin_home/src/arm_vision/msg/SegmentedClustersArray.msg
