+++ catkin_home/src/arm_vision/package.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>my_pcl_tutorial</name>
+  <name>arm_vision</name>
   <version>0.0.0</version>
-  <description>The my_pcl_tutorial package</description>
+  <description>The arm_vision package</description>
 
   <!-- One maintainer tag required, multiple allowed, one person per tag -->
   <!-- Example:  -->
@@ -19,7 +19,7 @@
   <!-- Url tags are optional, but multiple are allowed, one per tag -->
   <!-- Optional attribute type can be: website, bugtracker, or repository -->
   <!-- Example: -->
-  <!-- <url type="website">http://wiki.ros.org/my_pcl_tutorial</url> -->
+  <!-- <url type="website">http://wiki.ros.org/arm_vision</url> -->
 
 
   <!-- Author tags are optional, multiple are allowed, one per tag -->
